Split generic behavior tests
